Cleanup - moved apidya format to chw and renamed and modified to support Mr Nutz, Turrican 3 and Apidya.  Removed Eliminator as it is covered by chw. Removed Turrican III from factor 5
