+++ disk-analyse/formats
@@ -231,7 +231,7 @@ probe_amiga
     * amigados
 
 "Apidya"
-    * amigados apidya_a apidya_b apidya_c apidya_d
+    * amigados chw_2a chw_2b
 
 "Apprentice"
     * amigados apprentice
@@ -1357,9 +1357,9 @@ probe_amiga
 
 "Eliminator"
     0 amigados_unknown_length
-    1-79 eliminator_a eliminator_b
+    1-79 chw
     80 amigados_unknown_length
-    81-105 eliminator_a eliminator_b
+    81-105 chw
     106 amigados_unknown_length
     107-158 amigados_unknown_length
     * ignore
@@ -2810,6 +2810,9 @@ probe_amiga
 "Mr Blobby" = amigados
 
 "Mr DO! Run Run" = amigados_copylock
+
+"Mr Nutz - Hoppin' Mad"
+    * amigados_unknown_length chw_2a chw_2b
 
 # Track 0.1 does not appear to be formatted
 "Mugician"
@@ -4596,12 +4599,7 @@ probe_amiga
     * ignore
 
 "Turrican 3"
-    0 amigados
-    1 turrican_3b
-    2-42 turrican_3a
-    43 turrican_3c
-    44-159 turrican_3a  
-    * ignore
+    * amigados chw_2a chw_2b
 
 "Tusker"
     * amigados rainbird copylock
