+++ libdisk/format/amiga/chw.c
@@ -150,6 +150,136 @@ struct track_handler chw_handler = {
     .read_raw = chw_read_raw
 };
 
+
+/*
+ *
+ * Custom format as used on by the following games:
+ * 
+ *  Apidya
+ *  Turrican III
+ *  Mr. Nutz
+ *
+ * Written in 2022 by Keith Krellwitz 
+ *
+ * RAW TRACK LAYOUT:
+ * 
+ * TRKTYP_chw_2a
+ *  u16 (4489 Sync)
+ *  u16 padding - disk identifier
+ *  u32 dat[6144/4]
+ *  u32 Checksum
+ * 
+ * TRKTYP_chw_2b
+ *  u16 (4489 Sync)
+ *  u16 padding - disk identifier
+ *  u16 Track Number / 2
+ *  u32 dat[6656/4]
+ *  u32 Checksum
+ * 
+ * 
+ * Checksum: The sum of decoded data
+ *
+ */
+
+#include <libdisk/util.h>
+#include <private/disk.h>
+
+static void *chw_2_write_raw(
+    struct disk *d, unsigned int tracknr, struct stream *s)
+{
+    struct track_info *ti = &d->di->track[tracknr];
+
+    while (stream_next_bit(s) != -1) {
+
+        uint32_t raw[2], dat[ti->len/4+1], sum, csum;
+        uint16_t raw16[2], trk;
+        unsigned int i;
+        char *block;
+
+        /* sync */
+        if ((uint16_t)s->word != 0x4489)
+            continue;
+        ti->data_bitoff = s->index_offset_bc - 15;
+
+        /* disk identifier */
+        if (stream_next_bits(s, 16) == -1)
+            goto fail;
+        dat[ti->len/4] = (uint16_t)s->word;
+    
+        /* track number/2 only for track type TRKTYP_apidya_b */
+        if (ti->type == TRKTYP_chw_2b) {
+            if (stream_next_bytes(s, raw16, 4) == -1)
+                goto fail;
+            mfm_decode_bytes(bc_mfm_even_odd, 2, raw16, &trk);
+            if (be16toh(trk) != tracknr/2)
+                continue;
+        }
+
+        /* data */
+        for (i = sum = 0; i < ti->len/4; i++) {
+            if (stream_next_bytes(s, raw, 8) == -1)
+                goto fail;
+            mfm_decode_bytes(bc_mfm_even_odd, 4, raw, &dat[i]);
+            sum += be32toh(dat[i]);
+        }
+
+        /* checksum */
+        if (stream_next_bytes(s, raw, 8) == -1)
+            goto fail;
+        mfm_decode_bytes(bc_mfm_even_odd, 4, raw, &csum);
+
+        if (be32toh(csum) != sum)
+            continue;
+
+        stream_next_index(s);
+        ti->total_bits = (s->track_len_bc/100)*100+100;
+        block = memalloc(ti->len+4);
+        memcpy(block, dat, ti->len+4);
+        set_all_sectors_valid(ti);
+        return block;
+    }
+
+fail:
+    return NULL;
+}
+
+static void chw_2_read_raw(
+    struct disk *d, unsigned int tracknr, struct tbuf *tbuf)
+{
+    struct track_info *ti = &d->di->track[tracknr];
+    uint32_t *dat = (uint32_t *)ti->dat, sum;
+    unsigned int i;
+
+    tbuf_bits(tbuf, SPEED_AVG, bc_raw, 16, 0x4489);
+    tbuf_bits(tbuf, SPEED_AVG, bc_raw, 16, (uint16_t)dat[ti->len/4]);
+
+    if (ti->type == TRKTYP_chw_2b) {
+        tbuf_bits(tbuf, SPEED_AVG, bc_mfm_even_odd, 16, tracknr/2);    
+    }
+
+    for (i = sum = 0; i < ti->len/4; i++) {
+        tbuf_bits(tbuf, SPEED_AVG, bc_mfm_even_odd, 32, be32toh(dat[i]));
+        sum += be32toh(dat[i]);
+    }
+    tbuf_bits(tbuf, SPEED_AVG, bc_mfm_even_odd, 32, sum);
+}
+
+struct track_handler chw_2a_handler = {
+    .bytes_per_sector = 6144,
+    .nr_sectors = 1,
+    .write_raw = chw_2_write_raw,
+    .read_raw = chw_2_read_raw
+};
+
+struct track_handler chw_2b_handler = {
+    .bytes_per_sector = 6656,
+    .nr_sectors = 1,
+    .write_raw = chw_2_write_raw,
+    .read_raw = chw_2_read_raw
+};
+
+
+
 /*
  * Local variables:
  * mode: C
