+++ libdisk/include/libdisk/track_types.h
@@ -203,9 +203,6 @@ X(atari_st_720kb, "Atari ST 720kB")
 X(sales_curve, "Sales Curve")
 X(ninja_warriors, "Ninja Warriors")
 X(herndon_hls, "Herndon HLS")
-X(turrican_3a, "Turrican 3 A")
-X(turrican_3b, "Turrican 3 B")
-X(turrican_3c, "Turrican 3 C")
 X(head_over_heels, "Head Over Heels (Special FX)")
 X(f1, "F1")
 X(novagen, "Novagen")
@@ -245,10 +242,8 @@ X(actionware_protection, "Actionware Protection")
 X(hammerfist_a, "Hammerfist A")
 X(hammerfist_b, "Hammerfist B")
 X(hammerfist_c, "Hammerfist C")
-X(apidya_a, "Apidya A")
-X(apidya_b, "Apidya B")
-X(apidya_c, "Apidya C")
-X(apidya_d, "Apidya D")
+X(chw_2a, "Christian Weber 2A")
+X(chw_2b, "Christian Weber 2B")
 X(jupiters_masterdrive, "Jupiter's Masterdrive")
 X(celtic_legends_a, "Celtic Legends Disk 1")
 X(celtic_legends_b, "Celtic Legends Disk 2")
@@ -258,8 +253,6 @@ X(the_plague_c, "The Plague C")
 X(wayne_gretzky, "Wayne Gretzky")
 X(hellfire_attack, "Hellfire Attack Protection")
 X(deflektor, "Deflektor")
-X(eliminator_a, "Eliminator 198c")
-X(eliminator_b, "Eliminator 199c")
 X(sub_trade_a, "Subtrade: Return To Irata A")
 X(sub_trade_b, "Subtrade: Return To Irata B")
 X(ilyad, "Ilyad")
